Drop use of interrupts: do everything in a main routine.
Instead of reacting to a falling edge interrupt in SDA,
we wait for START condition in a main routine and do everything there.

This improves the latency from SDA falling to checking SCL is high
from between 3.25 and 4.25 instruction cycles (4.61us, out of spec)
to 2 instruction cycles (2.17us, well within spec)

It can also check that SCL was also high before the SDA falling edge
and so does not wrongly detect START conditions that are just other
devices' data streams.