Two changes: eliminate bcf on PORTs and make unstretch the last thing we do
Now that we are careful not to use bcf/bsf on PORT pins while talking
to the LCD module, the explicit clears of scl and sda when they become
outputs are no longer necessary.

We plan to hold scl low for the duration of the entire command/data
processing, so that the master need not use explicit delays after
data and commands. So make unstretching the clock the vary last thing
we do before returning from interrupt or going to check for a start
condition, in the hope that this gives us the maximum time to get it right.