Don't flick SDA low before setting it high
When sending a 0 bit, the code used to set SDA high then set it low again.
This is OK with the I2C spec since it was in the clock-low period,
but just setting it to the right value seems better, even if it takes
one instruction longer to do.