Fix commentary and remove the non-STERTCH_ON_SEND code.
There were tw version of i2c-sending code, one of which stretched the
clock on every bit and the otehr which just tried to go as fast as possible.
The fast-as possible only ran to 71kHz while the stretching runs at 91,
and the fast-as-possible version was 58 bytes bigger so had no advantages.